requirements.txt
# Telegram Bot
python-telegram-bot[job-queue]>=21.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# Google Gemini AI
google-genai>=1.0.0

# Scheduler
apscheduler>=3.10.0

# Database
aiosqlite>=0.19.0

# Environment
python-dotenv>=1.0.0

# Utilities
pytz>=2024.1
